+++ src/main/java/zd/zero/waifu/motivator/plugin/assets/AssetManager.kt
@@ -10,7 +10,7 @@ import org.apache.commons.io.IOUtils
 import org.apache.http.client.config.RequestConfig
 import org.apache.http.client.methods.HttpGet
 import org.apache.http.impl.client.HttpClients
-import zd.zero.waifu.motivator.plugin.remote.RestClient
+import zd.zero.waifu.motivator.plugin.tools.RestClient
 import zd.zero.waifu.motivator.plugin.tools.toOptional
 import java.io.IOException
 import java.nio.charset.Charset
@@ -33,7 +33,7 @@ enum class AssetCategory(val category: String) {
 }
 
 object AssetManager {
-    private const val ASSETS_SOURCE = "https://waifu-motivation-assets.s3.amazonaws.com"
+    private const val ASSETS_SOURCE = "https://waifu.assets.unthrottled.io"
 
     private val httpClient = HttpClients.custom()
         .setUserAgent(ApplicationInfoEx.getInstance().fullApplicationName)
