+++ src/main/java/zd/zero/waifu/motivator/plugin/tools/Extensions.kt
@@ -2,7 +2,7 @@ package zd.zero.waifu.motivator.plugin.tools
 
 import org.apache.commons.io.IOUtils
 import java.io.InputStream
-import java.util.*
+import java.util.Optional
 
 fun <T> T?.toOptional() = Optional.ofNullable(this)
